Add a Status property to File
There has been no way to determine if an upload succeeded. For example, if the
URL requested for the upload gives 404, the UploadMgr user still got a
"completed" event and no way to ascertain the failure.

This commit adds a new property, Status, to the File object. Since the File
object is given to the "completed" event, the user may now query this property
to ascertain success or failure.